refactor packages to move related files to its own package
