+++ src/functionalTest/java/uk/gov/hmcts/reform/wataskmonitor/MonitorTaskJobControllerTest.java
@@ -13,8 +13,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.wacaseeventhandler.TestUtility;
-import uk.gov.hmcts.reform.wataskmonitor.models.JobDetailName;
-import uk.gov.hmcts.reform.wataskmonitor.models.JobDetails;
+import uk.gov.hmcts.reform.wataskmonitor.models.jobs.JobDetailName;
+import uk.gov.hmcts.reform.wataskmonitor.models.jobs.JobDetail;
 import uk.gov.hmcts.reform.wataskmonitor.models.MonitorTaskJobReq;
 
 import static net.serenitybdd.rest.SerenityRest.given;
@@ -48,7 +48,7 @@ class MonitorTaskJobControllerTest {
         given()
             .contentType(APPLICATION_JSON_VALUE)
             .header("ServiceAuthorization", serviceToken)
-            .body(TestUtility.asJsonString(new MonitorTaskJobReq(new JobDetails(JobDetailName.CONFIGURATION))))
+            .body(TestUtility.asJsonString(new MonitorTaskJobReq(new JobDetail(JobDetailName.CONFIGURATION))))
             .when()
             .post("/monitor/tasks/jobs")
             .then()
