+++ src/main/java/uk/gov/hmcts/reform/wataskmonitor/models/jobs/JobDetailName.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.wataskmonitor.models;
+package uk.gov.hmcts.reform.wataskmonitor.models.jobs;
 
 public enum JobDetailName {
         TERMINATION, INITIATION, CONFIGURATION
