--- src/main/java/uk/gov/hmcts/reform/wataskmonitor/services/JobService.java
@@ -1,9 +1,0 @@
-package uk.gov.hmcts.reform.wataskmonitor.services;
-
-import uk.gov.hmcts.reform.wataskmonitor.models.JobDetailName;
-
-public interface JobService {
-    boolean canRun(JobDetailName jobDetailName);
-
-    void run();
-}
