+++ src/main/java/uk/gov/hmcts/reform/wataskmonitor/services/jobs/JobService.java
@@ -1,0 +1,9 @@
+package uk.gov.hmcts.reform.wataskmonitor.services.jobs;
+
+import uk.gov.hmcts.reform.wataskmonitor.models.jobs.JobDetailName;
+
+public interface JobService {
+    boolean canRun(JobDetailName jobDetailName);
+
+    void run();
+}
